# sit-event-api/src/event-registrations/event-registrations.controller.ts
import { Controller, Post, Param, Delete, HttpCode, Patch, Get, Query, Res } from '@nestjs/common';
import { EventRegistrationsService } from './event-registrations.service';
import type { Response } from 'express';
import { RegistrationStatus } from '../../generated/prisma';
import { 
  Roles, 
  AdminOnly, 
  EventOrganizerAccess, 
  AllRoleAccess,
  CurrentUser,
  type AuthenticatedUser,
  UserRole,
} from '../common';

@Controller('events')
export class EventRegistrationsController {
  constructor(private readonly eventRegistrationsService: EventRegistrationsService) {}

  @Post(':eventId/register')
  @AllRoleAccess()
  async registerUserToEvent(@Param('eventId') eventId: string, @CurrentUser() user: AuthenticatedUser) {
    return this.eventRegistrationsService.registerUserToEvent(eventId, user);
  }

  @Delete(':eventId/unregister')
  @HttpCode(204)
  @AllRoleAccess()
  async unregisterUserFromEvent(@Param('eventId') eventId: string, @CurrentUser() user: AuthenticatedUser) {
    return this.eventRegistrationsService.unregisterUserFromEvent(eventId, user);
  }

  @Get('registrations/me')
  @AllRoleAccess()
  async findMyRegistration(
    @CurrentUser() user: AuthenticatedUser,
  ) {
    return this.eventRegistrationsService.findMyRegistration(user);
  }

  @Delete('registrations/:registrationId/cancel')
  @HttpCode(204)
  @AdminOnly()
  async cancelRegistrationById(@Param('registrationId') registrationId: string) {
    return this.eventRegistrationsService.cancelRegistrationById(registrationId);
  }

  @Patch(':eventId/registrations/:registrationId')
  @AdminOnly()
  async changeAttendedStatusByRegistrationId(
    @Param('eventId') eventId: string,
    @Param('registrationId') registrationId: string,
  ) {
    return this.eventRegistrationsService.changeAttendedStatusByRegistrationId(
      eventId,
      registrationId,
      true,
    );
  }

  @Patch(':eventId/users/:userId')
  @AllRoleAccess()
  async changeAttendedStatusByUserId(
    @Param('eventId') eventId: string,
    @Param('userId') userId: string,
  ) {
    return this.eventRegistrationsService.changeAttendedStatusByUserId(
      eventId,
      userId,
      true,
    );
  }

  @Get(':eventId/open-qr/:userId')
  @AllRoleAccess() 
  async checkUserQrStatus(
    @Param('eventId') eventId: string,
    @Param('userId') userId: string,
  ) {
    return this.eventRegistrationsService.checkUserQrStatus(eventId, userId);
  }

  // =============================================
  // Get Registration Columns (Admin)
  // =============================================
  @Get(':eventId/registrations/columns')
  @AdminOnly()
  async getRegistrationColumns(@Param('eventId') eventId: string) {
    return this.eventRegistrationsService.getRegistrationColumns(eventId);
  }

  // =============================================
  // Get Pending Registrations (Admin)
  // =============================================
  @Get(':eventId/registrations/pending')
  @AdminOnly()
  async getPendingRegistrations(
    @Param('eventId') eventId: string,
    @Query('fields') fields?: string | string[],
    @Query('questionIds') questionIds?: string | string[],
  ) {
    // Normalize query params to string[]
    const parseArray = (input: string | string[] | undefined): string[] => {
      if (!input) return [];
      if (Array.isArray(input)) return input;
      return input.split(',').map((s) => s.trim());
    };

    const fieldsArray = parseArray(fields);
    const questionsArray = parseArray(questionIds);

    return this.eventRegistrationsService.getPendingRegistrations(
        eventId, 
        fieldsArray, 
        questionsArray
    );
  }

  // =============================================
  // Get Registrations by Status (Admin)
  // =============================================
  @Get(':eventId/registrations')
  @AdminOnly()
  async getRegistrationsByStatus(
    @Param('eventId') eventId: string,
    @Query('status') status?: RegistrationStatus,
  ) {
    return this.eventRegistrationsService.getRegistrationsByStatus(eventId, status);
  }

  // =============================================
  // Approve Registration (Admin)
  // =============================================
  @Patch(':eventId/registrations/:registrationId/approve')
  @AdminOnly()
  async approveRegistration(
    @Param('eventId') eventId: string,
    @Param('registrationId') registrationId: string,
  ) {
    return this.eventRegistrationsService.approveRegistration(eventId, registrationId);
  }

  // =============================================
  // Reject Registration (Admin)
  // =============================================
  @Patch(':eventId/registrations/:registrationId/reject')
  @AdminOnly()
  async rejectRegistration(
    @Param('eventId') eventId: string,
    @Param('registrationId') registrationId: string,
  ) {
    return this.eventRegistrationsService.rejectRegistration(eventId, registrationId);
  }

  // Check-in Event หลัก (Update)
  @Patch(':eventId/check-in/:userId')
  @AllRoleAccess() 
  async checkInUser(
    @Param('eventId') eventId: string,
    @Param('userId') userId: string,
  ) {
    return this.eventRegistrationsService.checkInUser(eventId, userId);
  }

  // [NEW] Check-in Sub-session
  @Patch(':eventId/check-in/:userId/:sessionId')
  @AllRoleAccess()
  async checkInUserSession(
    @Param('eventId') eventId: string,
    @Param('userId') userId: string,
    @Param('sessionId') sessionId: string,
  ) {
    return this.eventRegistrationsService.checkInUserSession(eventId, userId, sessionId);
  }

  // =============================================
  // Export Registrations (Admin)
  // =============================================
  @Get(':eventId/registrations/export')
  @AdminOnly()
  async exportRegistrations(
    @Param('eventId') eventId: string,
    @Res() res: Response,
  ) {
    return this.eventRegistrationsService.exportRegistrations(eventId, res);
  }
}