# sit-event-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event Management API",
    "description": "Complete API documentation for the event management system with authentication, events, registrations, and staff management.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/events": {
      "post": {
        "tags": ["Events"],
        "summary": "Create a new event (Admin/Organizer only)",
        "description": "Creates a new event with optional file uploads for thumbnail and images.",
        "security": [{ "JWT-auth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "thumbnail": { "type": "string", "format": "binary" },
                  "images": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" },
                    "maxItems": 5
                  },
                  "registrationOpenDate": { "type": "string", "format": "date-time" },
                  "registrationEndDate": { "type": "string", "format": "date-time" },
                  "eventStartDate": { "type": "string", "format": "date-time" },
                  "eventEndDate": { "type": "string", "format": "date-time" },
                  "targetAudience": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["EXTERNAL_STUDENT", "INTERNAL_STUDENT", "TEACHER", "PUBLIC"]
                    }
                  },
                  "tags": {
                    "type": "array",
                    "items": { "type": "string" }
                  },
                  "allowOnSiteRegister": { "type": "boolean" },
                  "activityHours": { "type": "integer" },
                  "invitation": { "type": "string" },
                  "website": { "type": "string", "format": "uri" },
                  "needWifi": { "type": "boolean" },
                  "certificateCriteria": { "type": "string" }
                },
                "required": ["name", "description", "registrationOpenDate", "registrationEndDate", "eventStartDate", "eventEndDate", "targetAudience", "tags"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Event created successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "get": {
        "tags": ["Events"],
        "summary": "Get all events (Public)",
        "description": "Returns a paginated list of all events.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer" },
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer" },
            "required": false
          }
        ],
        "responses": {
          "200": { "description": "List of all events" }
        }
      }
    },
...
    "/events/{eventId}/register": {
      "post": {
        "tags": ["Event Registrations"],
        "summary": "Register current user for an event",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "201": { "description": "Registered successfully" },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/events/{eventId}/unregister": {
      "delete": {
        "tags": ["Event Registrations"],
        "summary": "Unregister current user from an event",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Unregistered successfully" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Registration not found" }
        }
      }
    },
    "/events/registrations/me": {
      "get": {
        "tags": ["Event Registrations"],
        "summary": "Get my event registrations",
        "description": "Returns all event registrations for the current user.",
        "security": [{ "JWT-auth": [] }],
        "responses": {
          "200": { "description": "List of user's registrations" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/events/registrations/{registrationId}/cancel": {
      "delete": {
        "tags": ["Event Registrations"],
        "summary": "Cancel a registration by ID (Admin only)",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Registration canceled" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Registration not found" }
        }
      }
    },
    "/events/{eventId}/registrations": {
      "get": {
        "tags": ["Event Registrations"],
        "summary": "Get registrations by status (Admin only)",
        "description": "Returns all registrations for an event, optionally filtered by status.",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["PENDING", "APPROVED", "REJECTED"]
            },
            "description": "Filter registrations by status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of registrations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EventRegistration" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - Admin only" }
        }
      }
    },
    "/events/{eventId}/registrations/pending": {
      "get": {
        "tags": ["Event Registrations"],
        "summary": "Get pending registrations (Admin only)",
        "description": "Returns all pending registrations for an event waiting for approval.",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending registrations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EventRegistration" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - Admin only" }
        }
      }
    },
    "/events/{eventId}/registrations/{registrationId}/approve": {
      "patch": {
        "tags": ["Event Registrations"],
        "summary": "Approve a registration (Admin only)",
        "description": "Approves a pending registration, allowing the user to attend the event.",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Registration approved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventRegistration" }
              }
            }
          },
          "400": { "description": "Registration is already approved" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - Admin only" },
          "404": { "description": "Registration not found" }
        }
      }
    },
    "/events/{eventId}/registrations/{registrationId}/reject": {
      "patch": {
        "tags": ["Event Registrations"],
        "summary": "Reject a registration (Admin only)",
        "description": "Rejects a registration, preventing the user from attending the event.",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Registration rejected successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventRegistration" }
              }
            }
          },
          "400": { "description": "Registration is already rejected" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - Admin only" },
          "404": { "description": "Registration not found" }
        }
      }
    },
    "/events/{eventId}/registrations/{registrationId}": {
      "patch": {
        "tags": ["Event Registrations"],
        "summary": "Change attended status (Check-in) by Registration ID (Admin only)",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Status updated" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Registration not found" }
        }
      }
    },
    "/events/{eventId}/users/{userId}": {
      "patch": {
        "tags": ["Event Registrations"],
        "summary": "Change attended status (Check-in) by User ID",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Status updated" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User or registration not found" }
        }
      }
    },
    "/events/{eventId}/check-in/{userId}": {
      "patch": {
        "tags": ["Event Registrations"],
        "summary": "Check-in user to event",
        "description": "Marks a user as attended for an event.",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "User checked in successfully" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User or registration not found" }
        }
      }
    },
...
    "/forms/me": {
      "get": {
        "tags": ["Forms"],
        "summary": "Get all forms for the current user",
        "description": "Returns all forms from events that the user has attended, along with submission status.",
        "security": [{ "JWT-auth": [] }],
        "responses": {
          "200": {
            "description": "Returns the list of forms with submission status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyFormsResponseDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/events/{eventId}/forms": {
      "post": {
        "tags": ["Event Forms"],
        "summary": "Create a new form for an event",
        "description": "Creates a feedback/survey form for an event. Each event can only have one form.",
        "security": [{ "JWT-auth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Event ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEventFormDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Form successfully created" },
          "404": { "description": "Event not found" },
          "409": { "description": "Event already has a form" }
        }
      },
      "get": {
        "tags": ["Event Forms"],
        "summary": "Get the form for an event",
        "description": "Returns the form associated with an event, or null if no form exists.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Event ID"
          }
        ],
        "responses": {
          "200": { "description": "Returns the event form or null" },
          "404": { "description": "Event not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrentUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "username": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "roles": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
...
      "EventRegistration": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "eventId": { "type": "string", "format": "uuid" },
          "sessionId": { "type": "string", "format": "uuid", "nullable": true },
          "status": { 
            "type": "string", 
            "enum": ["PENDING", "APPROVED", "REJECTED"],
            "description": "Registration approval status"
          },
          "registeredAt": { "type": "string", "format": "date-time" },
          "approvedAt": { "type": "string", "format": "date-time", "nullable": true },
          "attended": { "type": "boolean" },
          "checkedInAt": { "type": "string", "format": "date-time", "nullable": true },
          "certificateIssued": { "type": "boolean" },
          "pointsEarned": { "type": "integer", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "user": { "$ref": "#/components/schemas/CurrentUser" },
          "event": { "$ref": "#/components/schemas/EventSummary" }
        }
      },
      "EventSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "eventStartDate": { "type": "string", "format": "date-time" },
          "eventEndDate": { "type": "string", "format": "date-time" }
        }
      },
      "CreateEventFormDto": {
        "type": "object",
        "properties": {
          "title": { 
            "type": "string",
            "example": "Event Feedback Form"
          },
          "description": { 
            "type": "string",
            "example": "Please share your feedback about the event"
          },
          "isActive": { 
            "type": "boolean",
            "default": false,
            "example": false
          },
          "fields": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CreateFormFieldDto" },
            "description": "Initial form fields to create with the form"
          }
        },
        "required": ["title"]
      },
...
      "CreateFormFieldDto": {
        "type": "object",
        "properties": {
          "question": { 
            "type": "string",
            "example": "How satisfied are you with the event?"
          },
          "fieldType": { 
            "type": "string",
            "enum": ["TEXT", "RATING_SCALE", "CHECKBOX", "RADIO"],
            "example": "RATING_SCALE"
          },
          "options": { 
            "type": "array",
            "items": { "type": "string" },
            "description": "Required for CHECKBOX and RADIO field types",
            "example": ["Option 1", "Option 2", "Option 3"]
          },
          "isRequired": { 
            "type": "boolean",
            "default": false,
            "example": true
          },
          "order": { 
            "type": "integer",
            "default": 0,
            "example": 1
          }
        },
        "required": ["question", "fieldType"]
      },
...
      "MyFormsResponseDto": {
        "type": "object",
        "properties": {
          "forms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserFormItemDto"
            }
          }
        }
      },
      "UserFormItemDto": {
        "type": "object",
        "properties": {
          "formId": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "eventName": {
            "type": "string",
            "example": "Tech Conference 2026"
          },
          "title": {
            "type": "string",
            "example": "Feedback Form"
          },
          "isSubmitted": {
            "type": "boolean",
            "example": true
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-01-30T10:00:00.000Z"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    }
  }
}